# docs/swagger.yaml
basePath: /
definitions:
...
  dto.FindWebhookResponse:
    properties:
      webhook:
        $ref: '#/definitions/models.InstanceWebhook'
    type: object
...
  dto.SetWebhookRequest:
    properties:
      webhook:
        $ref: '#/definitions/dto.SetWebhookRequestData'
    required:
    - webhook
    type: object
  dto.SetWebhookRequestData:
    properties:
      base64:
        type: boolean
      byEvents:
        type: boolean
      enabled:
        type: boolean
      events:
        items:
          type: string
        type: array
      headers:
        additionalProperties:
          type: string
        type: object
      url:
        type: string
    type: object
  dto.SetWebhookResponse:
    properties:
      webhook:
        $ref: '#/definitions/models.InstanceWebhook'
    type: object
...
  models.InstanceWebhook:
    properties:
      base64:
        type: boolean
      byEvents:
        type: boolean
      enabled:
        type: boolean
      events:
        items:
          type: string
        type: array
      headers:
        additionalProperties:
          type: string
        type: object
      url:
        type: string
    type: object
  utils.AuthenticationErrorResponse:
    properties:
      message:
        example: Unauthorized
        type: string
    type: object
  utils.HTTPErrorResponse:
    properties:
      error:
        description: |-
          Error is a legacy opaque value. Go's JSON encoder does not serialize an
          error message here, so clients should use Message and ErrorMessage.
        type: object
      errorMessage:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Verbeux AI
    url: https://github.com/verbeux-ai/whatsmiau
  description: Authenticated WhatsApp API compatible with Evolution API. The specification
    covers the complete application route surface, including canonical routes, compatibility
    aliases, calls and documentation delivery.
  license:
    name: MIT
  title: WhatsMiau API
  version: 0.3.2
paths:
...
  /v1/webhook/find/{instance}:
    get:
      operationId: getV1WebhookFindByInstance
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FindWebhookResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get an instance webhook
      tags:
      - Webhooks
  /v1/webhook/set/{instance}:
    post:
      consumes:
      - application/json
      description: Enables or updates the outbound webhook target. Call signaling
        is emitted only when the configured events include CALL; payloads never contain
        media, SDP, relay data, or call keys.
      operationId: postV1WebhookSetByInstance
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Webhook configuration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SetWebhookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SetWebhookResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Configure an instance webhook
      tags:
      - Webhooks
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: apikey
    type: apiKey
swagger: "2.0"